refactor(Comment): rename CommentDisplay to CommentCount and update imports across components
